requirements_colab.txt
astroid==3.1.0
black==24.3.0
bayesnf==0.1.3
chex==0.1.86
click==8.1.7
cloudpickle==3.0.0
dill==0.3.8
dm-tree==0.1.8
etils==1.7.0
exceptiongroup==1.2.1
execnet==2.1.1
fsspec==2024.3.1
iniconfig==2.0.0
isort==5.13.2
jaxtyping==0.2.28
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
optax==0.2.2
orbax-checkpoint==0.5.10
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
protobuf==5.26.1
Pygments==2.17.2
pyink==24.3.0
pylint==3.1.0
pytest==8.2.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
rich==13.7.1
scipy==1.13.0
six==1.16.0
tensorflow==2.18.0
tensorflow-probability==0.24.0
tf-keras==2.18.0
tomli==2.0.1
tomlkit==0.12.4
toolz==0.12.1
tqdm==4.66.2
typeguard==2.13.3
typing_extensions==4.11.0
tzdata==2024.1
